Reworked the quiz application from the ground up.
quiz is now contained in a module, fixing some of the problems I was having with private and public variables. Code has been further broken up into smaller functions/methods hopefully making it easier to update in the future.

The ideal of the current excercise is to build this with no jquery so I can practice my understanding of vanilla javascript, next up is learning jquery and refactoring this code for better browser compatability.